Moved styles to a dedicated file and changed method to deal with empty input values
The form styles have been moved to a dedicated file of its own. This is done so that it can be reused by other components that contain forms. Also rather than initializing the value of input elements as null in states, I have changed so that they are initialized as an empty string. I have also added a new method which makes sure to show error for empty values for a particular input only after the user has clicked off the input element